+++ src/hsreplaynet.d.ts
@@ -1,0 +1,10 @@
+export interface SingleCardDetailsPayload {
+	popularity_rank: number;
+	avg_copies_in_deck: number;
+	avg_turns_in_hand: number;
+	avg_turn_played: number;
+	keep_percentage: number;
+	winrate_in_opening_hand: number;
+	winrate_when_drawn: number;
+	winrate_when_played: number;
+}